The updated analysis page
