+++ client/src/App.js
@@ -35,14 +35,9 @@ function App() {
               <Route path="/overview" element={<Overview />} />
               <Route path="/discover/overview" element={<DiscoverOverview />} />
               <Route path="/discover/projects" element={<DiscoverProjects />} />
-              <Route path="/discover/teams" element={<DiscoverTeams />} />
-              <Route path="/discover/problem" element={<DiscoverProblem />} />
-              <Route path="/discover/lament" element={<DiscoverLament />} />
-              <Route path="/discern/overview" element={<DiscernOverview />} />
-              <Route path="/discern/brainstorm" element={<DiscernBrainstorm />} />
-              <Route path="/discern/analysis" element={<DiscernAnalysis />} />
-              <Route path="/discern/timeline" element={<DiscernTimeline />} />
-              <Route path="/testPageDatabase" element={<TestPageDatabase />} />
+              <Route path="/discover/teams" element={< DiscoverTeams />} />
+              <Route path="/discover/lament" element={<discoverLament />} />
+              <Route path="/discern/analysis" element={< DiscernAnalysis />} />
             </Routes>
           </div>
         </div>
