+++ client/src/pages/common/menuItems.jsx
@@ -11,10 +11,10 @@ export const discoverItems = [
 export const discernItems = [
   { label: 'Overview', path: '/discern/overview' },
   { label: 'Brainstorm', path: '/discern/brainstorm' },
-  { label: 'Analysis', path: '/discern/analysis' },
-  { label: 'Timeline', path: '/discern/timeline' }
+  { label: 'Analysis', path: '/discern/analysis' }
   ];
 export const developItems = [
+    
   ];
   
   export const demonstrateItems = [
