+++ client/src/pages/stylesheets/analysis.css
@@ -1,3 +1,6 @@
+.hero_analysis_page{
+  background-color: white;
+}
 .hero_analysis_header{
     border: 2px solid black;
     color: white;
@@ -57,6 +60,7 @@
 
 .solutions_input{
   margin: 2%;
+  border-bottom: 2px solid grey;
 }
 
 
@@ -75,9 +79,14 @@
 }
 
 .unassigned_Solutions{
-  margin: 2%;
+  margin: 2%; 
   display: flex;
   flex-direction: column;
+  margin-left: 300px;
+}
+
+.unassigned-Solutions h2{
+  margin-top: 5%;
 }
 
 .solution_explanation{
